requirements.txt
numpy
PyAudio
soundfile
onnxruntime
OpenAI
Pillow
